ruby: add support for version and versionFile using nixpkgs-ruby
Currently the ruby package can be manually changed by the user, but it
is often inconvenient to change it to a specific ruby version.
nixpkgs-ruby holds package definitions (and caches) all ruby versions.
Just like phps integration in devenv, this makes use of nixpkgs-ruby
whenever the version option is set.

In addition, since the ruby community is based mostly around rbenv,
support for setting the version using a versionFile (.ruby-version) is
also added. This makes use of nixpkgs-ruby's packageFromRubyVersionFile
helper function.

When when 2 or more of the version, versionFile or package properties
are set, a conflict will show up, because multiple sources are trying to
set the package property.