python: move venv and poetry scripts to separate files
Currently whenever an error occurs within the venv or poetry lines of
bash, an error occurs it is hard to track down these errors. This is
because the lines from enterShell are in a temporary file.

In addition, when tracking down this file, it is quite confusing seeing
the number of lines of different parts being there.

Having the code in separate files will clean up enterShell as well as
making errors easier to track down, as the errors now refer to these .sh
scripts with a line number.