python: rebuild venv when devenv-profile is updated
Currently whenever devenv-profile is updated, a new /nix/store path is
created for the python binary (and other dependencies). This will cause
problems when venv is already created against a different /nix/store
path. It'll use an older version when the old store path still exists
_or_ will fail when the old store path has been GCed.

This change makes sure to always recreate venv whenever devenv-profile
has changed. It checks this by storing a symlink to the devenv-profile
inside the venv and comparing this symlink each time devenv enters the
shell.