python: always refer to the poetry executable from devenv
Currently it can happen that poetry was installed within venv with a
different version than what is in devenv. This can lead to confusing
situations, especially when venv is being rebuild all of a sudden.

To circumvent this, we now explicitly refer to the poetry package
configused in devenv when executing `poetry install` inside of devenv.