python: poetry: always use pkgs.poetry
Previously poetry was overridden with the python version that was
supplied in config.languages.python.package.

However, the version of poetry that is in nixpkgs is not always
compatible with the version of python that is supplied in devenv.

Since Poetry should always be operating independently from
project-specific python version it should be fine to keep the
pkgs.poetry package as-is.