python: venv: use .venv instead of .devenv/state/venv
`.devenv/state/venv` was used to create virtual python environments. For
most projects this is fine, but for poetry there are some specific
exceptions where this isn't a good idea. `.venv` in the root of the
project is a common convention and poetry has explicit support for this
convention.

The problem where `.devenv/state/venv` is not a good idea is when the
python version in devenv does not match the python version-range in
pyproject.toml. Poetry will detect this and will try to create a new
virtualenv. It by default does this in `~/.cache/poetry/virtualenvs/`,
which results in a unexcpected shared virtual environment.

poetry has an option called
[virtualenvs.in-project](https://python-poetry.org/docs/configuration/#virtualenvsin-project)
which makes sure poetry stays within the project directory. This can
already be set globally in the user configuration, but setting this for
devenv is for a future commit.