use date from coreutils
Currently devenv uses the date executable that is available in PATH.
This means that under OSX a date executable is used that doesn't support
`N` formatting, which leads to `gc` directories being stored in seconds
instead of milliseconds.

This patch makes sure that date from coreutils is always used. This
should fix problems with conflicting gc files when multiple devenv's are
building in parallel.