devenv: use poetry instead of requirements.txt
Currently requirements.txt is used. It is installed in enterShell. This
has 2 downsides:

- Loading devenv shell takes a longer time, because pip install reruns
every time the devenv shell is entered.
- The dependencies that are installed are not locked. The latest version
is chosen, which will result in differences in tooling for different
people. devenv docs from the past might not be reproducible.

The biggest downside I have seen is that the files are larger in size.
Though I might be biased to using poetry vs pip directly.