postgres: add an `extensions` option
This is a convenience option to install a list of PostgreSQL extensions.
The set of available extensions is provided as an argument.

The description for the option contains a list of available extensions
generated from the current version of PostgreSQL.

You could already do this by modifying the package with `withPackages`
and then overriding `services.postgres.package`. The new option has two
advantages over the existing approach:

1. It's easier to find in the documentation.

2. Users unfamiliar with nixpkgs are unlikely to be aware of
   `pkgs.postgresql.withPackages` specifically, or the pattern in
   general.

The NixOS services calls this option `extraPlugins`. First of all,
that's not what PostgreSQL calls them. Second, you have to manually
source the extensions from `pkgs.postgresql.pkgs` and ensure that the
PostgreSQL version matches whatever is specified in
`services.postgres.package`.