Configure additional `rabbitmq` listeners
The rabbit server will listen for in-cluster communication and
local administration on a port calculated as:
`${cfg.port} + 20000`

This configures that port to use the same listener address as the
configuration option.

`epmd` is a companion process that is used to look up domain names
by `rabbitmq-server` and this also configures that process to listen
on the same listener address.