env: fix infinite recursion errors
The best way to solve this for now is to set these as part of
config.devenv.* namespace and reference these instead of from config.env
when setting env.

Also, when using flakes integration let's evaluate the whole config when
invoking `devenv up` instead of when building the devenv cli.

Fixes #580