[REFACTOR] Rework initialization script to be more indempotent
This change will "hide" the temporary server from the readiness checks
in addition to not listening on a TCP port. While initializing, the
readiness should be "not ready". It creates a temporary directory
inside of `$DEVENV_STATE` to hold the temporary UNIX socket. It also
overwrites `$PGHOST` to set the default for `psql` and restores the
old value when finished. It removes the temporary socket directory
explicitly and sets a `trap` to remove if there is an error.