[FEAT] Conditionally link devenv profile
When loading a shell, devenv checks .devenv/profile is

1. not a link
2. readlink reports someplace other than ${profile}

Then will remove whatever is .devenv/profile and relink to the current
active profile. When running a single command at a time, this is
usually not a problem, but if you attempt to run two commands at the
same time, then the race condition of one process removing a link while
another is creating the link will emit warnings to stderr.