Better (but not completely fixed) treatment of venv
- get rid of requirements check for recreation

- store .devenv_interpreter because the venv
  python binary resolves to the origin
  python of the wrapper, not the wrapper
  python

In practice, merging this commit will mean that

a) the invocation of pip install -r will
need to take care of any changed requirements.

b) the venv, once created, will never be recreated
unless the Python wrapper changes.  The Python
wrapper doesn't seem to change when you add
Python packages to your packages = list, I'm
not sure what the other circumstances might be.

In reality, this fix sucks because the venv is
not actually based on the wrapper.  The wrapper,
in fact, seems to never get used.  That will
need to be addressed in some other commit, though.