Add devenv-up into the shell environment
This brings devenv-up as a package which is brought into the shell
environment. `devenv-up` is now present in the $PATH and runs it
immediately and `devenv up` execs.

The main motivation here is when using a Flake directly. Previously
`devenv up` would run `nix build` on demand which is (1) fragile
since it makes assumptions around what the flake is actually called
and (2) doesn't register it as root.

By bringing the devenv-up into shell we solve both of these problems.
Be aware that we're not able to bring it into the _profile_ since this
leads to a infinite recursion: The devenv-up script is designed to run
"outside" of the devenv environment (since the process-compose.yaml
file sets the environment variables for you).

This commit doesn't change behavior in the Rust implementation, but
I believe we could now also tweak this to invoke `devenv-up` directly
instead of adding it as a root.