c: fix an issue where the wrong c/cpp compiler could be set up
We already add `stdenv`, which contains all the compilers and sets up
the `CC` and `CXX` env variables for each platform.
Adding `clang` or `gcc` on top can result in the wrong compilers being
set up for the current platform, e.g. gcc on macOS.
We saw this happen when building native extensions in ruby.