Make "postgres" process the top-level process
process-compose has functionality to show open ports
of the processes it manages:

https://github.com/F1bonacc1/process-compose/blob/2540374e305b0ac683ad88e06c637fe2fa7c0e5c/src/app/process.go#L736-L751

This doesn't work if the networked process is the child of the implicit `bash`
used by process-compose.

By running prepending `exec` we replace the implicit bash script
with the start-postgres script, which calls `exec postgres` internally.

This pattern should be enforced across the modules.