go: use non-versioned `buildGoModule` to support out-of-band compilers
If you set `languages.go.package` to a version of the go compiler that
doesn't yet exist in the `nixpkgs` input, you'll get an evaluation
error on `buildGoXXXModule` because the function for that specific
version doesn't exist.

I couldn't find anything in the docs preventing us from using the base
`buildGoModule` and override the `go` argument as we already do.