devenv: override temp dirs after nix develop
`nix develop` overrides all of the temp directories, which causes
problems for module evaluation after the shell, for example `devenv up`.

The temporary `nix-shell-<hash>` directory it creates is also unsafe to
use as it might be garbage collected at a later point.

The previous solution was to pick up the `DEVENV_RUNTIME` var, but that
is not safe for nested devenvs and does not solve the the tmpdir issue.