scripts: use lib.getExe to fetch the package binary by default
`scripts.<name>.binary` is now an optional override, in case
`lib.getExe` (i.e. `meta.mainProgram`) returns something that doesn't
work.

This fixes an issue in nixpkgs/master where the pname for bash is
changing to bash-interactive by default. Using the pname as the name of
the binary breaks scripts.