eval-cache: prevent cache invalidation due to timestamp precision issues
Use unix timestamps with second precision throughout.
This prevents cache invalidation stemming from multiple file paths that
differ in the timestamps only because of the truncation at the database
layer.

It'd be great to switch to microseconds at the database layer as well,
but we can't add the migration in a backwards-compatible way.