direnv: fix excessive reloads
When watching generated files, we need to take care not to touch the
files unless they change. direnv only compares the modification time,
not the file contents, so it's triggered every time we write these
files. Even worse, the vscode extension is happy to re-evaluate the
shell without waiting for the previous one to finish loading.