tasks: throttle the loop by waiting for updates
In #1844, the short-circuit that checks if there are no more tasks left
was moved after the notification waiter. This would cause the tasks to
hang in most cases, when there were no more tasks to send updates.

We still want to use the notification system even when there is no tty
to avoid wasteful loops.