Add readiness probes for elasticmq
By default the SQS-REST endpoint will listen on port 9324. If you change
this value, I give you the escape hatches to change where to ask the
server `/health`. The test in `examples/` has been updated to query
the server for the readiness endpoint, as well.

4 seconds may be a long time for a default installation, but that should
(hopefully) be sufficient for any customized installations to finish
booting.