tests: test flake-related edge-cases with git repos
Flakes are designed to work inside of git repos, which helps reduce the
cost of copying the source files to the Nix store.
This results in two separate code paths for devenv projects using git
and (usually ad-hoc) ones running sans git.
See #1990.

These tests are designed to catch regressions in the eval-cache
implementation stemming from these differences. If a Nix updates breaks
the path resolution, we will hopefully get a warning.