rust: fix toolchain creation with rust-overlay
In the switch from fenix to rust-overlay, the final toolchain was not
correctly patched. This resulted in runtime linker errors, where tools
like clippy were linked against the wrong librustc.

PR #1983 was an attempt at adjusting our APIs to better suit the
toolchain/profile design of rust-overlay. This PR retains the API
compatbility, at the cost of using private APIs.