devenv: fix issues loading .devenv internal files
#2094 broke the loading of the internal .devenv files.
The .devenv path needs to be a Nix path for this to work.

I've added comments explaining why it's set up the way it is.

Instead of just reverting, we now have a `devenv_dotfile_path`.
This lets us use `devenv_dotfile` for the absolute string path, matching `devenv_root`.
I think this is a bit clearer than the previous `devenv_dotfile_string` approach.