templates: make flake-parts use the same nixpkgs lib
flake-parts should use the same nixpkgs.lib we use for devenv to eval its modules.
This will prevent issues stemming from incompatible module system revisions.

We recently saw this with the v2 merge coherence check.
The module system added a boolean flag to every check and added an assertion checking for this flag.
The devenv modules were evaluated with a nixpkgs that did not have the flags, which then triggered the assertion on flake-part's side.