Clarify contributing-guide: local devenv input
I made the mistake of putting `devenv: url: path:<PATH-TO-DEVENV-SOURCE-CODE>?dir=src/modules` in the top-level of `devenv.yaml`. Instead in should go under the top-level key `inputs`, like `nixpkgs` already is.

I hope this documentation update will help the next person.

This diff also fixes so that bullet (4) is (4) and not (1) by indenting
code-block with 4 spaces.

Signed-off-by: Henrik Sjöstrand <henrik.sjostrand@tele2.com>