Introduced new option: services.mongodb.replication
Now, one will hardly run production environments and replication using devenv.
*However* when using MongoDB transactions their behaviour (read- and write concerns)
varies depending on if the MongoDB instance is set up for replication or not.

This patch gives the user the option to configure the devenv mongodb as a one-node replica-set.

When replication is enabled, MongoDB wants a `keyFile` to authenticate other nodes in the cluster.
This `keyFile` is created using openssl to make MongoDB happy.

Signed-off-by: Henrik Sjöstrand <henrik.sjostrand@tele2.com>