processes: prevent infinite recursion when conditionally defining processes
Fixes #2278.

This commit modifies the process module to always resolve and enable the correct process manager implementation.
This eliminates the possibility of triggering infinite recursion if a user uses mkIf to conditionally define processes based on an env var.

The module system would end up having to evaluate:

  config.processes -> config.env -> any mkIf'd env in the process manager -> config.process.managers.${implementation}.enable -> config.processes -> ...

Resulting in infinite recursion.